[REF] web: condition_tree: split file
The file condition_tree.js has become quite big. We split it in several
files so that the role of various parts are clearer. Along the way, we
have reworked a bit some functions, notably toAST.
