+++ addons/web/static/src/core/domain_selector/utils.js
@@ -1,8 +1,9 @@
-import { getDefaultValue } from "@web/core/tree_editor/tree_editor_value_editors";
 import { getDomainDisplayedOperators } from "@web/core/domain_selector/domain_selector_operator_editor";
-import { useService } from "@web/core/utils/hooks";
-import { domainFromTree, condition } from "@web/core/tree_editor/condition_tree";
+import { condition } from "@web/core/tree_editor/condition_tree";
+import { domainFromTree } from "@web/core/tree_editor/domain_from_tree";
+import { getDefaultValue } from "@web/core/tree_editor/tree_editor_value_editors";
 import { getDefaultPath } from "@web/core/tree_editor/utils";
+import { useService } from "@web/core/utils/hooks";
 
 export function getDefaultCondition(fieldDefs) {
     const defaultPath = getDefaultPath(fieldDefs);
