+++ addons/web/static/src/core/tree_editor/domain_contains_expressions.js
@@ -1,0 +1,33 @@
+import { Expression, isTree } from "./condition_tree";
+import { treeFromDomain } from "./tree_from_domain";
+
+function treeContainsExpressions(tree) {
+    if (tree.type === "condition") {
+        const { path, operator, value } = tree;
+        if (isTree(value) && treeContainsExpressions(value)) {
+            return true;
+        }
+        return [path, operator, value].some(
+            (v) =>
+                v instanceof Expression ||
+                (Array.isArray(v) && v.some((w) => w instanceof Expression))
+        );
+    }
+    for (const child of tree.children) {
+        if (treeContainsExpressions(child)) {
+            return true;
+        }
+    }
+    return false;
+}
+
+export function domainContainsExpressions(domain) {
+    let tree;
+    try {
+        tree = treeFromDomain(domain);
+    } catch {
+        return null;
+    }
+    // detect expressions in the domain tree, which we know is well-formed
+    return treeContainsExpressions(tree);
+}
