+++ addons/web/static/src/core/tree_editor/domain_from_tree.js
@@ -1,0 +1,7 @@
+import { constructDomainFromTree } from "./construct_domain_from_tree";
+import { eliminateVirtualOperators } from "./virtual_operators";
+
+export function domainFromTree(tree) {
+    const simplifiedTree = eliminateVirtualOperators(tree);
+    return constructDomainFromTree(simplifiedTree);
+}
