+++ addons/web/static/src/core/tree_editor/expression_from_tree.js
@@ -1,0 +1,7 @@
+import { constructExpressionFromTree } from "./construct_expression_from_tree";
+import { eliminateVirtualOperators } from "./virtual_operators";
+
+export function expressionFromTree(tree, options = {}) {
+    const simplifiedTree = eliminateVirtualOperators(tree);
+    return constructExpressionFromTree(simplifiedTree, options);
+}
