+++ addons/web/static/src/core/tree_editor/patterns/pattern.js
