+++ addons/web/static/src/core/tree_editor/tree_editor.js
@@ -2,13 +2,7 @@ import { Component, onWillStart, onWillUpdateProps } from "@odoo/owl";
 import { Dropdown } from "@web/core/dropdown/dropdown";
 import { DropdownItem } from "@web/core/dropdown/dropdown_item";
 import { useLoadFieldInfo } from "@web/core/model_field_selector/utils";
-import {
-    areEquivalentTrees,
-    cloneTree,
-    condition,
-    connector,
-    isTree,
-} from "@web/core/tree_editor/condition_tree";
+import { cloneTree, connector, isTree, TRUE_TREE } from "@web/core/tree_editor/condition_tree";
 import {
     getDefaultValue,
     getValueEditorInfo,
@@ -18,10 +12,10 @@ import {
     useMakeGetConditionDescription,
     useMakeGetFieldDef,
 } from "@web/core/tree_editor/utils";
+import { areEquivalentTrees } from "@web/core/tree_editor/virtual_operators";
 import { useService } from "@web/core/utils/hooks";
 import { shallowEqual } from "@web/core/utils/objects";
 
-const TRUE_TREE = condition(1, "=", 1);
 export class TreeEditor extends Component {
     static template = "web.TreeEditor";
     static components = {
