+++ addons/web/static/src/core/tree_editor/tree_from_domain.js
@@ -1,0 +1,7 @@
+import { constructTreeFromDomain } from "./construct_tree_from_domain";
+import { introduceVirtualOperators } from "./virtual_operators";
+
+export function treeFromDomain(domain, options = {}) {
+    const tree = constructTreeFromDomain(domain, options.distributeNot);
+    return introduceVirtualOperators(tree, options);
+}
