+++ addons/web/static/src/core/tree_editor/tree_from_expression.js
@@ -1,0 +1,7 @@
+import { constructTreeFromExpression } from "./construct_tree_from_expression";
+import { introduceVirtualOperators } from "./virtual_operators";
+
+export function treeFromExpression(expression, options = {}) {
+    const tree = constructTreeFromExpression(expression, options);
+    return introduceVirtualOperators(tree, options);
+}
