+++ addons/web/static/tests/core/tree_editor/condition_tree.test.js
@@ -1,22 +1,14 @@
 import { describe, expect, test } from "@odoo/hoot";
 
 import { makeMockEnv } from "@web/../tests/web_test_helpers";
-import { formatDomain } from "./condition_tree_editor_test_helpers";
 
 import { Domain } from "@web/core/domain";
 import { evaluateBooleanExpr } from "@web/core/py_js/py";
-import {
-    complexCondition,
-    condition,
-    connector,
-    constructDomain,
-    constructTree,
-    domainFromTree,
-    expression,
-    expressionFromTree,
-    treeFromDomain,
-    treeFromExpression,
-} from "@web/core/tree_editor/condition_tree";
+import { condition, expression } from "@web/core/tree_editor/condition_tree";
+import { domainFromTree } from "@web/core/tree_editor/domain_from_tree";
+import { expressionFromTree } from "@web/core/tree_editor/expression_from_tree";
+import { treeFromDomain } from "@web/core/tree_editor/tree_from_domain";
+import { treeFromExpression } from "@web/core/tree_editor/tree_from_expression";
 
 function expressionFromDomain(domain, options = {}) {
     const tree = treeFromDomain(domain, options);
@@ -29,75 +21,6 @@ function domainFromExpression(expression, options = {}) {
 }
 
 describe.current.tags("headless");
-
-test("domainFromTree", async () => {
-    await makeMockEnv();
-    const toTest = [
-        {
-            tree: condition("foo", "=", false),
-            result: `[("foo", "=", False)]`,
-        },
-        {
-            tree: condition("foo", "=", false, true),
-            result: `["!", ("foo", "=", False)]`,
-        },
-        {
-            tree: condition("foo", "=?", false),
-            result: `[("foo", "=?", False)]`,
-        },
-        {
-            tree: condition("foo", "=?", false, true),
-            result: `["!", ("foo", "=?", False)]`,
-        },
-        {
-            tree: condition("foo", "starts_with", "hello"),
-            result: `[("foo", "=ilike", "hello%")]`,
-        },
-        {
-            tree: condition("foo", "ends_with", "hello"),
-            result: `[("foo", "=ilike", "%hello")]`,
-        },
-        {
-            tree: condition("foo", "next", [1, "weeks", "date"]),
-            result: `["&", ("foo", ">=", context_today().strftime("%Y-%m-%d")), ("foo", "<=", (context_today() + relativedelta(weeks = 1)).strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "last", [1, "months", "date"]),
-            result: `["&", ("foo", ">=", (context_today() + relativedelta(months = -1)).strftime("%Y-%m-%d")), ("foo", "<=", context_today().strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "next", [1, "weeks", "datetime"]),
-            result: `["&", ("foo", ">=", datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")), ("foo", "<=", datetime.datetime.combine(context_today() + relativedelta(weeks = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S"))]`,
-        },
-        {
-            tree: condition("foo", "last", [1, "months", "datetime"]),
-            result: `["&", ("foo", ">=", datetime.datetime.combine(context_today() + relativedelta(months = -1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")), ("foo", "<=", datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S"))]`,
-        },
-        {
-            tree: condition("foo", "next", [1, "weeks", "date"], true),
-            result: `["!", "&", ("foo", ">=", context_today().strftime("%Y-%m-%d")), ("foo", "<=", (context_today() + relativedelta(weeks = 1)).strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "last", [expression("a"), "weeks", "date"], true),
-            result: `["!", "&", ("foo", ">=", (context_today() + relativedelta(weeks = a)).strftime("%Y-%m-%d")), ("foo", "<=", context_today().strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "next", [1, "b", "date"], true),
-            result: `["!", "&", ("foo", ">=", context_today().strftime("%Y-%m-%d")), ("foo", "<=", (context_today() + relativedelta(b = 1)).strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "not_next", [1, "weeks", "date"]),
-            result: `["|", ("foo", "<", context_today().strftime("%Y-%m-%d")), ("foo", ">", (context_today() + relativedelta(weeks = 1)).strftime("%Y-%m-%d"))]`,
-        },
-        {
-            tree: condition("foo", "not_next", [1, "weeks", "date"], true),
-            result: `["!", "|", ("foo", "<", context_today().strftime("%Y-%m-%d")), ("foo", ">", (context_today() + relativedelta(weeks = 1)).strftime("%Y-%m-%d"))]`,
-        },
-    ];
-    for (const { tree, result } of toTest) {
-        expect(domainFromTree(tree).replace(/[\s\n]+/g, "")).toBe(result.replace(/[\s\n]+/g, ""));
-    }
-});
 
 test("domainFromTree . treeFromDomain", async () => {
     await makeMockEnv();
@@ -300,429 +223,6 @@ test("domainFromExpression", () => {
     for (const { expression, result, extraOptions } of toTest) {
         const o = { ...options, ...extraOptions };
         expect(domainFromExpression(expression, o)).toBe(result);
-    }
-});
-
-test("expressionFromTree", () => {
-    const options = {
-        getFieldDef: (name) => {
-            if (["foo", "bar"].includes(name)) {
-                return { type: "any" }; // any field
-            }
-            if (["foo_ids", "bar_ids"].includes(name)) {
-                return { type: "many2many" };
-            }
-            return null;
-        },
-    };
-    const toTest = [
-        {
-            expressionTree: condition("foo", "=", false),
-            result: `not foo`,
-        },
-        {
-            expressionTree: condition("foo", "=", false, true),
-            result: `foo`,
-        },
-        {
-            expressionTree: condition("foo", "!=", false),
-            result: `foo`,
-        },
-        {
-            expressionTree: condition("foo", "!=", false, true),
-            result: `not foo`,
-        },
-        {
-            expressionTree: condition("y", "=", false),
-            result: `not "y"`,
-        },
-        {
-            expressionTree: condition("foo", "between", [1, 3]),
-            result: `foo >= 1 and foo <= 3`,
-        },
-        {
-            expressionTree: condition("foo", "between", [1, expression("uid")], true),
-            result: `not ( foo >= 1 and foo <= uid )`,
-        },
-        {
-            expressionTree: condition("foo", "next", [1, "weeks", "date"]),
-            result: `foo >= context_today().strftime("%Y-%m-%d") and foo <= (context_today() + relativedelta(weeks = 1)).strftime("%Y-%m-%d")`,
-        },
-        {
-            expressionTree: condition("foo", "last", [1, "weeks", "date"]),
-            result: `foo >= (context_today() + relativedelta(weeks = -1)).strftime("%Y-%m-%d") and foo <= context_today().strftime("%Y-%m-%d")`,
-        },
-        {
-            expressionTree: condition("foo", "next", [1, "months", "datetime"]),
-            result: `foo >= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and foo <= datetime.datetime.combine(context_today() + relativedelta(months = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-        },
-        {
-            expressionTree: condition("foo", "last", [1, "months", "datetime"]),
-            result: `foo >= datetime.datetime.combine(context_today() + relativedelta(months = -1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and foo <= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-        },
-        {
-            expressionTree: condition("foo", "last", [expression("a"), "months", "datetime"]),
-            result: `foo >= datetime.datetime.combine(context_today() + relativedelta(months = a), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and foo <= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-        },
-        {
-            expressionTree: condition("foo", "last", [1, "b", "datetime"]),
-            result: `foo >= datetime.datetime.combine(context_today() + relativedelta(b = -1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and foo <= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-        },
-        {
-            expressionTree: complexCondition("uid"),
-            result: `uid`,
-        },
-        {
-            expressionTree: condition("foo_ids", "in", []),
-            result: `set(foo_ids).intersection([])`,
-        },
-        {
-            expressionTree: condition("foo_ids", "in", [1]),
-            result: `set(foo_ids).intersection([1])`,
-        },
-        {
-            expressionTree: condition("foo_ids", "in", 1),
-            result: `set(foo_ids).intersection([1])`,
-        },
-        {
-            expressionTree: condition("foo", "in", []),
-            result: `foo in []`,
-        },
-        {
-            expressionTree: condition(expression("expr"), "in", []),
-            result: `expr in []`,
-        },
-        {
-            expressionTree: condition("foo", "in", [1]),
-            result: `foo in [1]`,
-        },
-        {
-            expressionTree: condition("foo", "in", 1),
-            result: `foo in [1]`,
-        },
-        {
-            expressionTree: condition("foo", "in", expression("expr")),
-            result: `foo in expr`,
-        },
-        {
-            expressionTree: condition("foo_ids", "in", expression("expr")),
-            result: `set(foo_ids).intersection(expr)`,
-        },
-        {
-            expressionTree: condition("y", "in", []),
-            result: `"y" in []`,
-        },
-        {
-            expressionTree: condition("y", "in", [1]),
-            result: `"y" in [1]`,
-        },
-        {
-            expressionTree: condition("y", "in", 1),
-            result: `"y" in [1]`,
-        },
-        {
-            expressionTree: condition("foo_ids", "not in", []),
-            result: `not set(foo_ids).intersection([])`,
-        },
-        {
-            expressionTree: condition("foo_ids", "not in", [1]),
-            result: `not set(foo_ids).intersection([1])`,
-        },
-        {
-            expressionTree: condition("foo_ids", "not in", 1),
-            result: `not set(foo_ids).intersection([1])`,
-        },
-        {
-            expressionTree: condition("foo", "not in", []),
-            result: `foo not in []`,
-        },
-        {
-            expressionTree: condition("foo", "not in", [1]),
-            result: `foo not in [1]`,
-        },
-        {
-            expressionTree: condition("foo", "not in", 1),
-            result: `foo not in [1]`,
-        },
-        {
-            expressionTree: condition("y", "not in", []),
-            result: `"y" not in []`,
-        },
-        {
-            expressionTree: condition("y", "not in", [1]),
-            result: `"y" not in [1]`,
-        },
-        {
-            expressionTree: condition("y", "not in", 1),
-            result: `"y" not in [1]`,
-        },
-    ];
-    for (const { expressionTree, result, extraOptions } of toTest) {
-        const o = { ...options, ...extraOptions };
-        expect(expressionFromTree(expressionTree, o)).toBe(result);
-    }
-});
-
-test("treeFromExpression", () => {
-    const options = {
-        getFieldDef: (name) => {
-            if (["foo", "bar"].includes(name)) {
-                return { type: "any" }; // any field
-            }
-            if (["foo_ids", "bar_ids"].includes(name)) {
-                return { type: "many2many" };
-            }
-            if (name === "date_field") {
-                return { type: "date" };
-            }
-            if (name === "datetime_field") {
-                return { type: "datetime" };
-            }
-            return null;
-        },
-    };
-    const toTest = [
-        {
-            expression: `not foo`,
-            result: condition("foo", "not_set", false),
-        },
-        {
-            expression: `foo == False`,
-            result: condition("foo", "not_set", false),
-        },
-        {
-            expression: `foo`,
-            result: condition("foo", "set", false),
-        },
-        {
-            expression: `foo == True`,
-            result: condition("foo", "=", true),
-        },
-        {
-            expression: `foo is True`,
-            result: complexCondition(`foo is True`),
-        },
-        {
-            expression: `not (foo == False)`,
-            result: condition("foo", "set", false),
-        },
-        {
-            expression: `not (not foo)`,
-            result: condition("foo", "set", false),
-        },
-        {
-            expression: `foo >= 1 and foo <= 3`,
-            result: condition("foo", "between", [1, 3]),
-        },
-        {
-            expression: `foo >= 1 and foo <= uid`,
-            result: condition("foo", "between", [1, expression("uid")]),
-        },
-        {
-            expression: `foo < 1 or foo > 3`,
-            result: condition("foo", "not_between", [1, 3]),
-        },
-        {
-            expression: `date_field >= context_today().strftime("%Y-%m-%d") and date_field <= (context_today() + relativedelta(years = 1)).strftime("%Y-%m-%d")`,
-            result: condition("date_field", "next", [1, "years", "date"]),
-        },
-        {
-            expression: `date_field < context_today().strftime("%Y-%m-%d") or date_field > (context_today() + relativedelta(years = 1)).strftime("%Y-%m-%d")`,
-            result: condition("date_field", "not_next", [1, "years", "date"]),
-        },
-        {
-            expression: `datetime_field >= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and datetime_field <= datetime.datetime.combine(context_today() + relativedelta(years = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-            result: condition("datetime_field", "next", [1, "years", "datetime"]),
-        },
-        {
-            // Case where the <= is first: this is not changed to a between, and so not changed to a within either
-            expression: `datetime_field <= datetime.datetime.combine(context_today() + relativedelta(years = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and datetime_field >= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-            result: connector("&", [
-                condition(
-                    "datetime_field",
-                    "<=",
-                    expression(
-                        `datetime.datetime.combine(context_today() + relativedelta(years = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`
-                    )
-                ),
-                condition(
-                    "datetime_field",
-                    ">=",
-                    expression(
-                        `datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`
-                    )
-                ),
-            ]),
-        },
-        {
-            expression: `datetime_field >= datetime.datetime.combine(context_today() + relativedelta(years = 1), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S") and datetime_field <= datetime.datetime.combine(context_today(), datetime.time(0, 0, 0)).to_utc().strftime("%Y-%m-%d %H:%M:%S")`,
-            result: condition("datetime_field", "last", [-1, "years", "datetime"]),
-        },
-        {
-            expression: `(date_field >= context_today().strftime("%Y-%m-%d") and date_field <= (context_today() + relativedelta(years = 1)).strftime("%Y-%m-%d")) and (date_field >= context_today().strftime("%Y-%m-%d") and date_field <= (context_today() + relativedelta(years = 2)).strftime("%Y-%m-%d"))`,
-            result: connector("&", [
-                condition("date_field", "next", [1, "years", "date"]),
-                condition("date_field", "next", [2, "years", "date"]),
-            ]),
-        },
-        {
-            expression: `(date_field >= context_today().strftime("%Y-%m-%d") and date_field <= (context_today() + relativedelta(years = 1)).strftime("%Y-%m-%d")) or (date_field >= context_today().strftime("%Y-%m-%d") and date_field <= (context_today() + relativedelta(years = 2)).strftime("%Y-%m-%d"))`,
-            result: connector("|", [
-                condition("date_field", "next", [1, "years", "date"]),
-                condition("date_field", "next", [2, "years", "date"]),
-            ]),
-        },
-        {
-            expression: `foo >= 1 if bar else foo <= uid`,
-            result: connector("|", [
-                connector("&", [condition("bar", "set", false), condition("foo", ">=", 1)]),
-                connector("&", [
-                    condition("bar", "not_set", false),
-                    condition("foo", "<=", expression("uid")),
-                ]),
-            ]),
-        },
-        {
-            expression: `context.get('toto')`,
-            result: complexCondition(`context.get("toto")`),
-        },
-        {
-            expression: `not context.get('toto')`,
-            result: complexCondition(`not context.get("toto")`),
-        },
-        {
-            expression: `foo >= 1 if context.get('toto') else bar == 42`,
-            result: connector("|", [
-                connector("&", [
-                    complexCondition(`context.get("toto")`),
-                    condition("foo", ">=", 1),
-                ]),
-                connector("&", [
-                    complexCondition(`not context.get("toto")`),
-                    condition("bar", "=", 42),
-                ]),
-            ]),
-        },
-        {
-            expression: `set()`,
-            result: complexCondition(`set()`),
-        },
-        {
-            expression: `set([1, 2])`,
-            result: complexCondition(`set([1, 2])`),
-        },
-        {
-            expression: `set(foo_ids).intersection([1, 2])`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set(foo_ids).intersection(set([1, 2]))`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set(foo_ids).intersection(set((1, 2)))`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set(foo_ids).intersection("ab")`,
-            result: complexCondition(`set(foo_ids).intersection("ab")`),
-        },
-        {
-            expression: `set([1, 2]).intersection(foo_ids)`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set(set([1, 2])).intersection(foo_ids)`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set((1, 2)).intersection(foo_ids)`,
-            result: condition("foo_ids", "in", [1, 2]),
-        },
-        {
-            expression: `set("ab").intersection(foo_ids)`,
-            result: complexCondition(`set("ab").intersection(foo_ids)`),
-        },
-        {
-            expression: `set([2, 3]).intersection([1, 2])`,
-            result: complexCondition(`set([2, 3]).intersection([1, 2])`),
-        },
-        {
-            expression: `set(foo_ids).intersection(bar_ids)`,
-            result: complexCondition(`set(foo_ids).intersection(bar_ids)`),
-        },
-        {
-            expression: `set().intersection(foo_ids)`,
-            result: condition(0, "=", 1),
-        },
-        {
-            expression: `set(foo_ids).intersection()`,
-            result: condition("foo_ids", "set", false),
-        },
-        {
-            expression: `not set().intersection(foo_ids)`,
-            result: condition(1, "=", 1),
-        },
-        {
-            expression: `not set(foo_ids).intersection()`,
-            result: condition("foo_ids", "not_set", false),
-        },
-        {
-            expression: `not set(foo_ids).intersection([1, 2])`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set(foo_ids).intersection(set([1, 2]))`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set(foo_ids).intersection(set((1, 2)))`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set(foo_ids).intersection("ab")`,
-            result: complexCondition(`not set(foo_ids).intersection("ab")`),
-        },
-        {
-            expression: `not set([1, 2]).intersection(foo_ids)`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set(set([1, 2])).intersection(foo_ids)`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set((1, 2)).intersection(foo_ids)`,
-            result: condition("foo_ids", "not in", [1, 2]),
-        },
-        {
-            expression: `not set("ab").intersection(foo_ids)`,
-            result: complexCondition(`not set("ab").intersection(foo_ids)`),
-        },
-        {
-            expression: `not set([2, 3]).intersection([1, 2])`,
-            result: complexCondition(`not set([2, 3]).intersection([1, 2])`),
-        },
-        {
-            expression: `not set(foo_ids).intersection(bar_ids)`,
-            result: complexCondition(`not set(foo_ids).intersection(bar_ids)`),
-        },
-        {
-            expression: `set(foo_ids).difference([1, 2])`,
-            result: complexCondition(`set(foo_ids).difference([1, 2])`),
-        },
-        {
-            expression: `set(foo_ids).union([1, 2])`,
-            result: complexCondition(`set(foo_ids).union([1, 2])`),
-        },
-        {
-            expression: `expr in []`,
-            result: complexCondition(`expr in []`),
-        },
-    ];
-    for (const { expression, result, extraOptions } of toTest) {
-        const o = { ...options, ...extraOptions };
-        expect(treeFromExpression(expression, o)).toEqual(result);
     }
 });
 
@@ -1042,78 +542,5 @@ test("evaluation . expressionFromTree = contains . domainFromTree", () => {
         expect(evaluateBooleanExpr(expressionFromTree(tree, options), record)).toBe(
             new Domain(domainFromTree(tree)).contains(record)
         );
-    }
-});
-
-test("constructTree", async () => {
-    const toTest = [
-        { tree: connector("&"), domain: `[]` },
-        { tree: condition(0, "=", 1), domain: `[(0, "=", 1)]` },
-        { tree: condition(1, "=", 1), domain: `[(1, "=", 1)]` },
-        { tree: condition(0, "=", 1, true), domain: `["!", (0, "=", 1)]` },
-        { tree: condition(1, "=", 1, true), domain: `["!", (1, "=", 1)]` },
-    ];
-    for (const { tree, domain } of toTest) {
-        expect(constructTree(domain)).toEqual(tree);
-    }
-});
-
-test("constructDomain", async () => {
-    const toTest = [
-        { tree: connector("&"), domain: `[]` },
-        { tree: connector("&", [], true), domain: `[(0, "=", 1)]` },
-        { tree: connector("|"), domain: `[(0, "=", 1)]` },
-        { tree: connector("|", [], true), domain: `[(1, "=", 1)]` },
-        { tree: condition(1, "=", 1), domain: `[(1, "=", 1)]` },
-        { tree: condition(0, "=", 1), domain: `[(0, "=", 1)]` },
-        { tree: condition(1, "=", 1, true), domain: `["!", (1, "=", 1)]` },
-        { tree: condition(0, "=", 1, true), domain: `["!", (0, "=", 1)]` },
-        { tree: connector("|", [connector("|")]), domain: `[(0, "=", 1)]` },
-        { tree: connector("|", [connector("&")]), domain: `[(1, "=", 1)]` },
-        { tree: connector("&", [connector("&")]), domain: `[(1, "=", 1)]` },
-        { tree: connector("&", [connector("|")]), domain: `[(0, "=", 1)]` },
-        {
-            tree: connector("|", [connector("|"), condition("id", "=", 1)]),
-            domain: `["|", (0, "=", 1), ("id", "=", 1)]`,
-        },
-        {
-            tree: connector("|", [connector("&"), condition("id", "=", 1)]),
-            domain: `["|", (1, "=", 1), ("id", "=", 1)]`,
-        },
-        {
-            tree: connector("&", [connector("&"), condition("id", "=", 1)]),
-            domain: `["&", (1, "=", 1), ("id", "=", 1)]`,
-        },
-        {
-            tree: connector("&", [connector("|"), condition("id", "=", 1)]),
-            domain: `["&", (0, "=", 1), ("id", "=", 1)]`,
-        },
-        {
-            tree: condition("id", "=", 1),
-            domain: `[("id", "=", 1)]`,
-        },
-        {
-            tree: condition("m2m", "any", connector("&")),
-            domain: `[("m2m", "any", [])]`,
-        },
-        {
-            tree: condition("m2m", "any", connector("&", [connector("&")])),
-            domain: `[("m2m", "any", [(1, "=", 1)])]`,
-        },
-        {
-            tree: condition("m2m", "any", connector("&", [connector("|")])),
-            domain: `[("m2m", "any", [(0, "=", 1)])]`,
-        },
-        {
-            tree: condition("m2m", "any", connector("|", [connector("|")])),
-            domain: `[("m2m", "any", [(0, "=", 1)])]`,
-        },
-        {
-            tree: condition("m2m", "any", connector("|", [connector("&")])),
-            domain: `[("m2m", "any", [(1, "=", 1)])]`,
-        },
-    ];
-    for (const { tree, domain } of toTest) {
-        expect(constructDomain(tree)).toBe(formatDomain(domain));
     }
 });
